Change the .NET data type used for FoxPro float data type to decimal

+++ DotNetDBF/DBFWriter.cs
@@ -438,7 +438,7 @@ namespace DotNetDBF
 
                         if (objectArray[j] != null && objectArray[j] != DBNull.Value)
                         {
-                            var tDouble = Convert.ToDouble(objectArray[j]);
+                            var tDouble = Convert.ToDecimal(objectArray[j]);
                             dataOutput.Write(
                                 Utils.NumericFormating(
                                     tDouble,
